Moved schematron-related UI into separate plugin project
