+++ io/features/eu.esdihumboldt.hale.io.feature.core.ui/feature.xml
@@ -72,4 +72,11 @@
          version="0.0.0"
          unpack="false"/>
 
+   <plugin
+         id="eu.esdihumboldt.hale.io.schematron.ui"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
 </feature>
