+++ io/plugins/eu.esdihumboldt.hale.io.schematron.ui/.settings/org.eclipse.jdt.launching.prefs
@@ -1,0 +1,4 @@
+#Created from default preferences 16.12.2016 12:55:44
+#Fri Dec 16 12:55:44 CET 2016
+eclipse.preferences.version=1
+org.eclipse.jdt.launching.PREF_STRICTLY_COMPATIBLE_JRE_NOT_AVAILABLE=ignore
