+++ io/plugins/eu.esdihumboldt.hale.io.schematron.ui/build.properties
@@ -1,0 +1,6 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               src/,\
+               plugin.xml
