+++ io/plugins/eu.esdihumboldt.hale.io.schematron.ui/plugin.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="eu.esdihumboldt.hale.ui.io.config">
+      <configDialog
+            class="eu.esdihumboldt.hale.io.schematron.ui.SchematronValidatorConfigurationDialog">
+         <provider
+               ref="eu.esdihumboldt.hale.io.schematron.validator">
+         </provider>
+      </configDialog>
+   </extension>
+
+</plugin>
