+++ ui/plugins/eu.esdihumboldt.hale.ui/META-INF/MANIFEST.MF
@@ -84,6 +84,7 @@ Import-Package: com.google.common.base;version="1.6.0",
  eu.esdihumboldt.hale.common.schema.model.impl,
  eu.esdihumboldt.hale.common.scripting.scripts.groovy,
  eu.esdihumboldt.hale.common.service.helper.population,
+ eu.esdihumboldt.hale.io.schematron.ui,
  eu.esdihumboldt.hale.io.xsd.constraint,
  eu.esdihumboldt.hale.ui.common,
  eu.esdihumboldt.hale.ui.common.definition,
