+++ ui/plugins/eu.esdihumboldt.hale.ui/src/eu/esdihumboldt/hale/ui/io/config/ConfigurationDialogExtension.java
@@ -126,7 +126,7 @@ public class ConfigurationDialogExtension
 	}
 
 	/**
-	 * Factory for {@link AbstractConfigurationPage}s based on an
+	 * Factory for {@link AbstractConfigurationDialog}s based on an
 	 * {@link IConfigurationElement}
 	 */
 	private static class ConfigurationDialogFactoryImpl
